requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.4.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.8.6
asttokens==2.2.1
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
bandit==1.7.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
black==22.8.0
bleach==6.0.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
contourpy==1.1.1
coverage==7.2.1
cryptography==39.0.2
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-cytoscape==1.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-cli==0.17.5
dbus-python==1.2.16
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.4.0
distlib==0.3.8
distro==1.8.0
docker==6.0.1
doit==0.33.1
dpath==2.1.4
dulwich==0.21.3
dvc==2.24.0
dvc-data==0.4.0
dvc-http==2.19.1
dvc-objects==0.2.0
dvc-render==0.0.9
dvc-task==0.1.2
dvclive==0.11.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
feyn==3.2.0
filelock==3.13.4
flake8==3.9.1
Flask==2.2.3
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.51.0
frozenlist==1.3.3
fsspec==2023.3.0
funcy==1.18
future==0.18.3
gevent==24.2.1
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
grandalf==0.6
greenlet==2.0.2
gunicorn==20.1.0
hydra-core==1.3.2
identify==2.5.35
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
interpret==0.6.1
interpret-core==0.6.1
ipykernel==6.29.4
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
isort==5.12.0
itsdangerous==2.1.2
jedi==0.17.2
Jinja2==3.1.2
joblib==1.4.0
json5==0.9.11
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-http-over-ws==0.0.8
jupyter-server==1.23.6
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab==3.4.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.20.0
jupyterlab_widgets==3.0.10
kaleido==0.2.1
kiwisolver==1.4.5
kombu==5.2.4
lark==1.1.9
lazy-object-proxy==1.9.0
llvmlite==0.41.1
loguru==0.5.3
Mako==1.2.4
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.5
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
mergedeep==1.3.4
missingno==0.5.2
mistune==2.0.5
mkdocs==1.4.2
mkdocs-git-revision-date-plugin==0.3.1
mkdocs-material==8.1.9
mkdocs-material-extensions==1.1.1
mlflow==1.21.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nanotime==0.5.2
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nodeenv==1.8.0
notebook==6.5.3
notebook_shim==0.2.2
numba==0.58.1
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.7.1
pathspec==0.9.0
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.21.0
pluggy==0.13.1
polars==0.19.19
pre-commit==3.5.0
prometheus-client==0.16.0
prometheus-flask-exporter==0.22.3
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==13.0.0
pycodestyle==2.7.0
pycparser==2.21
pydot==1.4.2
pyflakes==2.3.1
pygit2==1.11.1
Pygments==2.14.0
PyGObject==3.36.0
pygtrie==2.5.0
pyinotify==0.9.6
PyJWT==2.6.0
pylint==2.11.1
pymdown-extensions==9.10
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==6.1.1
pytest-cov==2.11.1
python-dateutil==2.8.2
python-dotenv==0.19.1
python-editor==1.0.4
pytz==2022.7.1
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==25.0.1
qtconsole==5.5.1
QtPy==2.4.1
querystring-parser==1.2.4
requests==2.28.2
retrying==1.3.4
rich==13.3.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SALib==1.4.8
scikit-learn==1.3.2
scipy==1.10.1
scmrepo==0.0.25
seaborn==0.13.2
Send2Trash==1.8.0
shap==0.44.1
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
skimpy==0.0.11
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.6
sqlparse==0.4.3
ssh-import-id==5.10
stack-data==0.6.2
stevedore==5.0.0
svgwrite==1.4.3
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.4.0
tinycss2==1.2.1
tokenize-rt==4.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typeguard==4.1.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.25.3
voluptuous==0.13.1
watchdog==2.3.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.10
wrapt==1.13.3
yarl==1.8.2
zc.lockfile==3.0.post1
zipp==3.15.0
zope.event==5.0
zope.interface==6.3
# specific gpg requirements
sympy==1.11.1
setuptools==65.4.0
torch==1.13.1
functorch==1.13.1
sympytorch==0.1.4
stopit==1.1.2
pybind11
